rename rest service
